rename: study 패키지명을 step1으로 변경
