+++ src/test/java/step1/StringTests.java
@@ -1,4 +1,4 @@
-package study;
+package step1;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
